Get URLs to Available Radolan Files

Get URLs to available radolan files on the DWD FTP server below this base address: ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany. The user can choose between daily records and hourly records that are located at different paths on the server. The paths to the files are not read from the FTP server but generated, given the knowledge of where the files should reside and how they are expected to be named.

get_radolan_urls(start_daily = "2006-10", start_hourly = "2005-06",
  end_daily = format(format = "%Y-%m", lubridate::rollback(Sys.Date(),
  roll_to_first = TRUE)), end_hourly = end_daily)

Arguments

start_daily
month string (yyyy-mm) of first daily record. Default: "2006-10"

start_hourly
month string (yyyy-mm) of first hourly record. Default: "2005-06"

end_daily
month string (yyyy-mm) of last daily record. Defaults to the current month.

end_hourly
month string (yyyy-mm) of last hourly records. Defaults to end_daily.

Value

list with "daily_historical_urls" and "hourly_historical_urls"

Examples

# Get all expected URLs
urls <- kwb.dwd::get_radolan_urls()

# Show the first URLs of each category (daily, hourly)
head(urls$daily_historical_urls)

#> [1] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/daily/radolan/historical/2006/SF-200610.tar.gz"
#> [2] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/daily/radolan/historical/2006/SF-200611.tar.gz"
#> [3] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/daily/radolan/historical/2006/SF-200612.tar.gz"
#> [4] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/daily/radolan/historical/2007/SF-200701.tar.gz"
#> [5] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/daily/radolan/historical/2007/SF-200702.tar.gz"
#> [6] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/daily/radolan/historical/2007/SF-200703.tar.gz"

head(urls$hourly_historical_urls)

#> [1] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/hourly/radolan/historical/bin/2005/RW-200506.tar.gz"
#> [2] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/hourly/radolan/historical/bin/2005/RW-200507.tar.gz"
#> [3] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/hourly/radolan/historical/bin/2005/RW-200508.tar.gz"
#> [4] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/hourly/radolan/historical/bin/2005/RW-200509.tar.gz"
#> [5] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/hourly/radolan/historical/bin/2005/RW-200510.tar.gz"
#> [6] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/hourly/radolan/historical/bin/2005/RW-200511.tar.gz"

# Set the start and end months
urls <- kwb.dwd::get_radolan_urls(
  start_daily = "2017-01", end_daily = "2017-03"
)

urls$daily_historical_urls

#> [1] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/daily/radolan/historical/2017/SF201701.tar.gz"
#> [2] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/daily/radolan/historical/2017/SF201702.tar.gz"
#> [3] "ftp://ftp-cdc.dwd.de/pub/CDC/grids_germany/daily/radolan/historical/2017/SF201703.tar.gz"